Move OpExtInst validation into validate_extensions.cpp
Fixes #2123
