+++ source/CMakeLists.txt
@@ -306,7 +306,6 @@ set(SPIRV_SOURCES
   ${CMAKE_CURRENT_SOURCE_DIR}/val/validate_decorations.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/val/validate_derivatives.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/val/validate_extensions.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/val/validate_ext_inst.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/val/validate_execution_limitations.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/val/validate_function.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/val/validate_id.cpp
